better organization with header files for more efficient build
