+++ 2048.h
@@ -1,0 +1,3 @@
+#include "global.h"
+#include "game.h"
+#include "menu.h"
