+++ menu.cpp
@@ -1,67 +1,62 @@
-#pragma once
+#include "menu.h"
 
-#include <iostream>
-#include <iomanip>
-#include <cstdlib>
+void Menu::startMenu(int err) {
 
-#include "color.hpp"
+        clearScreen();
 
-#ifdef _WIN32
+        drawAscii();
+        std::cout << bold_on << "  Welcome to " << blue << "2048!" << def << bold_off; endl(2);
+        std::cout << "          1. Play a New Game"; endl();
+        std::cout << "          2. View Highscores and Statistics"; endl(2);
 
-    void getInput(char &c) {
-        std::cin >> c;
-    }
+        input(err);
 
-#else
-    #include <unistd.h>
-    #include <termios.h>
-
-    char getch() {
-        char buf = 0;
-        struct termios old = {0};
-        if (tcgetattr(0, &old) < 0)
-            perror("tcsetattr()");
-        old.c_lflag &= ~ICANON;
-        old.c_lflag &= ~ECHO;
-        old.c_cc[VMIN] = 1;
-        old.c_cc[VTIME] = 0;
-        if (tcsetattr(0, TCSANOW, &old) < 0)
-            perror("tcsetattr ICANON");
-        if (read(0, &buf, 1) < 0)
-            perror("read()");
-        old.c_lflag |= ICANON;
-        old.c_lflag |= ECHO;
-        if (tcsetattr(0, TCSADRAIN, &old) < 0)
-            perror("tcsetattr ~ICANON");
-        return (buf);
-    }
-
-    void getInput(char &c) {
-        c = getch();
-    }
-
-#endif
-
-typedef unsigned long long ull;
-
-void endl(int n = 1) {
-
-    for (int i = 0; i < n; i++) {
-        std::cout << std::endl;
-    }
-
-};
-
-void clearScreen() {
-    system("clear");
-};
-
-std::ostream &bold_on(std::ostream &os) {
-    return os << "\e[1m";
 }
 
-std::ostream &bold_off(std::ostream &os) {
-    return os << "\e[0m";
+void Menu::input(int err) {
+
+        if (err) {
+                std::cout << red << "  Invalid input. Please try again." << def; endl(2);
+        }
+
+        std::cout << "  Enter Choice: ";
+        char c;
+        std::cin >> c;
+
+        if (std::cin.eof())
+        {
+            std::cout << std::endl;
+            exit(EXIT_SUCCESS);
+  }
+
+        switch(c) {
+                case '1':
+                        startGame();
+                        break;
+                case '2':
+                        showScores();
+                        break;
+                default:
+                        startMenu(1);
+                        break;
+        }
+
+}
+
+void Menu::startGame() {
+
+        Game g;
+        g.startGame();
+
+}
+
+void Menu::showScores() {
+
+        Scoreboard s;
+        s.printScore();
+        s.printStats();
+        return;
+
 }
 
 void drawAscii() {
